Add helpers to template
