+++ packages/tools/src/helpers.ts
@@ -18,26 +18,8 @@ if (!process.env.PROJECT_CWD) {
 }
 export const ROOT = process.env.PROJECT_CWD!;
 
-export const IMPORTS_KEY = '#imports-injection-placeholder';
-export const TYPES_KEY = '#types-injection-placeholder';
-export const CONFIGS_KEY = '#configs-injection-placeholder';
-
 const execFileP = promisify(execFile);
 const maxBuffer = 1024 * 1024;
 
 export const execute = (cmd: string, args: string[], cwd: string) =>
     execFileP(cmd, args, { maxBuffer, cwd, encoding: 'utf-8' });
-
-export const injectIntoString = (content: string, mark: string, injection: string) => {
-    // Find the mark
-    const contentArray = content.split('\n');
-    const index = contentArray.findIndex((line) => line.includes(mark));
-
-    // Inject the new content
-    if (index === -1) {
-        throw new Error(`Could not find the mark ${green(mark)} in the content.`);
-    }
-    contentArray.splice(index, 0, injection);
-
-    return contentArray.join('\n');
-};
